requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart==0.0.9
pandas>=2.0.0
openpyxl>=3.1.0
weasyprint>=60.0
python-dotenv==1.0.0
